pyproject.toml
[project]
name = "bot-reminder"
version = "0.1.0"
description = "Бот-напоминалка"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.26.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "dataclasses>=0.8",
    "freezegun>=1.5.5",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "tzlocal>=5.3.1",
    "virtualenv>=21.2.0",
    "vkbottle>=4.7.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
